+++ src/ecs/components/society_agent.hpp
@@ -36,6 +36,6 @@ struct SocietyAgent
   SocialClass social_class;
   Metier metiers;
   State state = State::Idle;
-  int time_to_next_action = 0;
+  double time_to_next_action = 0.0;
 };
 }  // namespace dl
