+++ src/scenes/world_creation.hpp
@@ -20,7 +20,7 @@ class WorldCreation : public Scene
   WorldCreation(const std::string& scene_key, Camera& camera);
 
   void load() override;
-  void update(const uint32_t delta, SetSceneFunction set_scene) override;
+  void update(const double delta, SetSceneFunction set_scene) override;
   void render(Renderer& renderer) override;
   /* void screenshot(tcod::Context& context, TCOD_Console& console, const std::string& filename) override; */
